experiments/local/requirements_lightweight.txt
# 軽量展示用バックエンド依存関係 - 完全無料版

# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart>=0.0.6

# PyTorch (CPU版 - 軽量)
torch>=2.0.0+cpu
torchvision>=0.15.0+cpu
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# System Monitoring (Optional)
psutil>=5.9.0

# Development Tools (Optional)
requests>=2.31.0

# Notes:
# - All packages are free and open source
# - CPU-only PyTorch for lightweight deployment
# - No GPU dependencies required
# - Total download size: ~500MB
# - Commercial use allowed under respective licenses